Fix castling error
Fixed an error where a king could still castle when a pawn attacked
one of the castle squares. In this present inplementation, pawns only
add squares to their available_moves when an actual piece is in the
attacked square. And that is not the case when checking if an empty
castle square is attacked.